Made persiststate module
