+++ counter/build.gradle
@@ -16,6 +16,7 @@ android {
 
 dependencies {
     implementation project(':mvrx')
+    implementation project(':mvrx-persiststate')
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlinVersion"
     implementation "androidx.appcompat:appcompat:$appCompatVersion"
     implementation "androidx.core:core-ktx:$androidXCoreKtxVersion"
