+++ mvrx-persiststate/gradle.properties
@@ -1,0 +1,4 @@
+POM_NAME=MvRx Persist State
+POM_ARTIFACT_ID=mvrx-persiststate
+POM_PACKAGING=aar
+GROUP=com.airbnb.android
