+++ mvrx/src/main/kotlin/com/airbnb/mvrx/MvRxMutabilityHelper.kt
@@ -37,23 +37,26 @@ internal fun KClass<*>.assertImmutability() {
         }
     }
 
-    java.declaredFields.forEach { prop ->
-        when {
-            !Modifier.isFinal(prop.modifiers) -> "State property ${prop.name} must be a val, not a var."
-            prop.isSubtype(ArrayList::class) -> "You cannot use ArrayList for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
-            prop.isSubtype(SparseArray::class) -> "You cannot use SparseArray for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
-            prop.isSubtype(LongSparseArray::class) -> "You cannot use LongSparseArray for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
-            prop.isSubtype(SparseArrayCompat::class) -> "You cannot use SparseArrayCompat for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
-            prop.isSubtype(ArrayMap::class) -> "You cannot use ArrayMap for ${prop.name}.\n$IMMUTABLE_MAP_MESSAGE"
-            Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT &&
-                prop.isSubtype(android.util.ArrayMap::class) -> "You cannot use ArrayMap for ${prop.name}.\n$IMMUTABLE_MAP_MESSAGE"
-            prop.isSubtype(HashMap::class) -> "You cannot use HashMap for ${prop.name}.\n$IMMUTABLE_MAP_MESSAGE"
-            prop.isSubtype(Function::class, KCallable::class) -> {
-                "You cannot use functions inside MvRx state. Only pure data should be represented: ${prop.name}"
-            }
-            else -> null
-        }?.let { throw IllegalArgumentException(it) }
-    }
+    java.declaredFields
+        // During tests, jacoco can add a transient field called jacocoData.
+        .filter { !Modifier.isTransient(it.modifiers) }
+        .forEach { prop ->
+            when {
+                !Modifier.isFinal(prop.modifiers) -> "State property ${prop.name} must be a val, not a var."
+                prop.isSubtype(ArrayList::class) -> "You cannot use ArrayList for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
+                prop.isSubtype(SparseArray::class) -> "You cannot use SparseArray for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
+                prop.isSubtype(LongSparseArray::class) -> "You cannot use LongSparseArray for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
+                prop.isSubtype(SparseArrayCompat::class) -> "You cannot use SparseArrayCompat for ${prop.name}.\n$IMMUTABLE_LIST_MESSAGE"
+                prop.isSubtype(ArrayMap::class) -> "You cannot use ArrayMap for ${prop.name}.\n$IMMUTABLE_MAP_MESSAGE"
+                Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT &&
+                    prop.isSubtype(android.util.ArrayMap::class) -> "You cannot use ArrayMap for ${prop.name}.\n$IMMUTABLE_MAP_MESSAGE"
+                prop.isSubtype(HashMap::class) -> "You cannot use HashMap for ${prop.name}.\n$IMMUTABLE_MAP_MESSAGE"
+                prop.isSubtype(Function::class, KCallable::class) -> {
+                    "You cannot use functions inside MvRx state. Only pure data should be represented: ${prop.name}"
+                }
+                else -> null
+            }?.let { throw IllegalArgumentException(it) }
+        }
 }
 
 /**
