+++ settings.gradle
@@ -5,3 +5,4 @@ include ':testing'
 include ':dogs'
 include ':counter'
 include ':hellodagger'
+include ':mvrx-persiststate'
